+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # 0.4
 
 - Changed attribute definitions to be an array instead of an object and added deprecation message. 
+
+# 0.5
+
+- Chain expression `.toJS()` now returns `action` instead of `actions` array if there is only one action.
+  Both are still valid to parse.
+- Misc bug fixes